Don't cast the line encoding baud rate to a uint16_t before performing the baud rate calculation in the Benito project.

+++ Projects/Benito/Benito.c
@@ -229,7 +229,7 @@ void EVENT_CDC_Device_LineEncodingChanged(USB_ClassInfo_CDC_Device_t* const CDCI
 	UCSR1A = (1 << U2X1);
 	UCSR1B = ((1 << RXCIE1) | (1 << TXEN1) | (1 << RXEN1));
 	UCSR1C = ConfigMask;	
-	UBRR1  = SERIAL_2X_UBBRVAL((uint16_t)CDCInterfaceInfo->State.LineEncoding.BaudRateBPS);
+	UBRR1  = SERIAL_2X_UBBRVAL(CDCInterfaceInfo->State.LineEncoding.BaudRateBPS);
 }
 
 /** ISR to manage the reception of data from the serial port, placing received bytes into a circular buffer
